[mod] move extract_text, extract_url to searx.utils
