+++ searx/engines/acgsou.py
@@ -11,8 +11,7 @@
 
 from urllib.parse import urlencode
 from lxml import html
-from searx.engines.xpath import extract_text
-from searx.utils import get_torrent_size, int_or_zero
+from searx.utils import extract_text, get_torrent_size, int_or_zero
 
 # engine dependent config
 categories = ['files', 'images', 'videos', 'music']
