+++ searx/engines/apkmirror.py
@@ -11,7 +11,7 @@
 
 from urllib.parse import urlencode
 from lxml import html
-from searx.engines.xpath import extract_text
+from searx.utils import extract_text
 
 
 # engine dependent config
