+++ searx/engines/btdigg.py
@@ -13,8 +13,7 @@
 from lxml import html
 from operator import itemgetter
 from urllib.parse import quote, urljoin
-from searx.engines.xpath import extract_text
-from searx.utils import get_torrent_size
+from searx.utils import extract_text, get_torrent_size
 
 # engine dependent config
 categories = ['videos', 'music', 'files']
