+++ searx/engines/deviantart.py
@@ -15,7 +15,7 @@
 from lxml import html
 import re
 from urllib.parse import urlencode
-from searx.engines.xpath import extract_text
+from searx.utils import extract_text
 
 
 # engine dependent config
