+++ searx/engines/doku.py
@@ -11,8 +11,7 @@
 
 from urllib.parse import urlencode
 from lxml.html import fromstring
-from searx.engines.xpath import extract_text
-from searx.utils import eval_xpath
+from searx.utils import extract_text, eval_xpath
 
 # engine dependent config
 categories = ['general']  # TODO , 'images', 'music', 'videos', 'files'
