+++ searx/engines/duden.py
@@ -11,8 +11,7 @@
 from lxml import html, etree
 import re
 from urllib.parse import quote, urljoin
-from searx.engines.xpath import extract_text
-from searx.utils import eval_xpath
+from searx.utils import extract_text, eval_xpath
 from searx import logger
 
 categories = ['general']
