+++ searx/engines/framalibre.py
@@ -13,7 +13,7 @@
 from html import escape
 from urllib.parse import urljoin, urlencode
 from lxml import html
-from searx.engines.xpath import extract_text
+from searx.utils import extract_text
 
 # engine dependent config
 categories = ['it']
