+++ searx/engines/google_videos.py
@@ -14,7 +14,7 @@ from datetime import date, timedelta
 from json import loads
 from urllib.parse import urlencode
 from lxml import html
-from searx.engines.xpath import extract_text
+from searx.utils import extract_text
 import re
 
 # engine dependent config
