+++ searx/engines/ina.py
@@ -16,7 +16,7 @@ from urllib.parse import urlencode
 from lxml import html
 from dateutil import parser
 from html.parser import HTMLParser
-from searx.engines.xpath import extract_text
+from searx.utils import extract_text
 
 
 # engine dependent config
