+++ searx/engines/stackoverflow.py
@@ -12,7 +12,7 @@
 
 from urllib.parse import urlencode, urljoin
 from lxml import html
-from searx.engines.xpath import extract_text
+from searx.utils import extract_text
 
 # engine dependent config
 categories = ['it']
