+++ searx/engines/tokyotoshokan.py
@@ -13,9 +13,8 @@
 import re
 from urllib.parse import urlencode
 from lxml import html
-from searx.engines.xpath import extract_text
 from datetime import datetime
-from searx.utils import get_torrent_size, int_or_zero
+from searx.utils import extract_text, get_torrent_size, int_or_zero
 
 # engine dependent config
 categories = ['files', 'videos', 'music']
