+++ searx/engines/twitter.py
@@ -15,7 +15,7 @@
 from urllib.parse import urlencode, urljoin
 from lxml import html
 from datetime import datetime
-from searx.engines.xpath import extract_text
+from searx.utils import extract_text
 
 # engine dependent config
 categories = ['social media']
