+++ searx/engines/wikidata.py
@@ -13,9 +13,8 @@
 
 from searx import logger
 from searx.poolrequests import get
-from searx.engines.xpath import extract_text
 from searx.engines.wikipedia import _fetch_supported_languages, supported_languages_url
-from searx.utils import match_language, eval_xpath
+from searx.utils import extract_text, match_language, eval_xpath
 
 from urllib.parse import urlencode
 from json import loads
