+++ searx/engines/youtube_noapi.py
@@ -11,8 +11,7 @@
 from functools import reduce
 from json import loads
 from urllib.parse import quote_plus
-from searx.engines.xpath import extract_text
-from searx.utils import list_get
+from searx.utils import extract_text, list_get
 
 # engine dependent config
 categories = ['videos', 'music']
